pyproject.toml
[tool.poetry]
name = "msc_thesis"
version = "0.1.0"
description = "This is the virtual environment used for my MSc thesis"
authors = ["Philip Hartout <philip.hartout@protonmail.com>"]
license = "BSD-3 Clause License"

[tool.poetry.dependencies]
python = "~3.9"
pandas = "^1.4.0"
pre-commit = "^2.17.0"
matplotlib = "^3.5.1"
pytest = "^6.2.5"
isort = "^5.10.1"
jupyter = "^1.0.0"
mypy = "^0.931"
networkx = "^2.6.3"
numpy = "^1.22.1"
pandoc = "^2.0.1"
PyYAML = "^6.0"
scikit-learn = "^1.0.2"
tqdm = "^4.62.3"
pyprojroot = "^0.2.0"
giotto-tda = "^0.5.1"
grakel = "^0.1.8"
biopython = "^1.79"
fastwlk = "^0.2.14"
seaborn = "^0.11.2"
kaleido = "0.2.1"
nx = "^0.0.0"
black = "^22.3.0"
hydra-core = "^1.1.1"
plotly = "^5.6.0"
imageio = "^2.16.1"
natsort = "^8.1.0"
fair-esm = "^0.4.0"
torch = "^1.11.0"
scipy = "^1.8.0"
umap-learn = "^0.5.3"
Brotli = "^1.0.9"

[tool.poetry.dev-dependencies]
black = {version = "^22.1.0", allow-prereleases = true}
proteinmetrics = { path = "./src/", develop = true }
snakeviz = "^2.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
target-version = ['py38']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 79
src_paths = ["src", "tests"]
float_to_top = true

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
